+++ src/TRSConversion/Parse/Utils.hs
@@ -1,6 +1,3 @@
-{-# LANGUAGE ImportQualifiedPost #-}
-{-# LANGUAGE OverloadedStrings #-}
-
 -- |
 -- Module      : TRSConversion.Parse.Utils
 -- Description : Utils for parsing with Megaparsec
@@ -14,10 +11,8 @@ module TRSConversion.Parse.Utils
 where
 
 import Data.Text (Text)
-import Data.Void (Void)
 import Text.Megaparsec (Parsec)
 
 -- | Type alias for a 'Megaparsec' parser which uses error handler of type 'Void' and
 -- takes an input of type 'Text'.
-type Parser = Parsec Void Text
-
+type Parser e = Parsec e Text